requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
python-dotenv==1.0.0
groq==0.4.2
Werkzeug==2.0.1
gunicorn==20.1.0
python-jose==3.3.0
bcrypt==4.0.1
pyrebase4==4.8.0
requests==2.32.3
urllib3==1.26.20
gcloud==0.18.3
oauth2client==4.1.3
pycryptodome==3.22.0
requests-toolbelt==0.10.1
python-jwt==2.0.1
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.1.8
langchain
langchain-groq
markdown

